requirements-test.txt
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27.0
pyjwt>=2.8.0
cryptography>=42.0.0
